requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
Jinja2==3.1.3
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.5
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
bcrypt==4.1.2
stripe==7.11.0
SQLAlchemy==2.0.27
alembic==1.11.3
python-dotenv==1.0.1
email-validator==2.1.0.post1
typing-extensions>=4.5.0
Flask-Mail==0.9.1
PyMySQL==1.1.0
Flask-WTF==1.2.1
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==41.0.7
distro==1.9.0
dnspython==2.7.0
google==3.0.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud==0.34.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-oauth2-tool==0.0.3
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.1
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.9.0
Mako==1.3.10
oauth2client==4.1.3
oauthlib==3.2.2

packaging==25.0
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
tqdm==4.67.1
typing-inspection==0.4.0
uritemplate==4.1.1
urllib3==2.4.0
WTForms==3.2.1
musicai-sdk==1.0.2 